+++ resources/views/modules/product/same-products.blade.php
@@ -1,23 +1,11 @@
 @vite('resources/css/modules/product/same-products.css')
+@vite('resources/css/modules/product/mini-product-card.css')
 
 @if ($same_products->isNotEmpty())
 <p id="same-product-title">Check out related products:</p>
 <div id="same-products">
-    @foreach ($same_products as $same_product)
-        <a class="same-product__card" href="{{ route('product', ['id' => $same_product->id]) }}">
-            <img src="{{ $same_product->image_url }}" alt="product-image">
-            <p class="same-product__data"><span>Name:</span> {{ $same_product->name }}</p>
-            <p class="same-product__data"><span>Price:</span> {{ $same_product->price }}$</p>
-            <p class="same-product__filters">
-                <span class="same-product__filter-tag">Filters:</span>
-                @foreach ($filters as $filter)
-                    <span class="{{ $filter->filter_name }}">
-                        {{ $filter->filter_name }}
-                    </span>
-                @endforeach
-            </p>
-            <p class="same-product__more">View more</p>
-        </a>
+    @foreach ($same_products as $product)
+        @include('modules.product.mini-product-card')
     @endforeach
         <a href="{{ route('products') }}" class="same-products__more-text">
             <span>View more Products</span>